doc: fix syntax for a couple of surfacer options
Noticed that the surfacer doc showed a couple of options that did not match the `key: value` syntax as the rest of the doc. I tested this by setting these options exactly as I saw them in the doc, getting an error on the cloudprober pod upon startup, fixing the syntax, and watching the pod start up without fail.